Move tests outside of module
