+++ .github/workflows/test-all-commits.yml
@@ -18,7 +18,7 @@ jobs:
        run: |
          sudo apt update
          sudo apt install python3-pip -y
-         pip3 install -r pt_os_web_portal/tests/requirements.txt
+         pip3 install -r tests/requirements.txt
 
      - name: Run Python tests and generate coverage report
        run: |
