+++ setup.cfg
@@ -67,4 +67,4 @@ norecursedirs =
     debian
     docs
     frontend
-    pt_os_web_portal/tests/old
+    tests/old
