+++ tests/data/wifi_manager_data.py
