+++ tests/mocked_system_folder/locale.gen
