+++ tests/old/test_finalise.py
@@ -2,6 +2,7 @@ from os import path
 from shlex import split
 
 from flask import json
+
 from tests.data.finalise_data import (
     available_space,
     available_space_out,
