+++ tests/old/test_keyboard.py
@@ -1,4 +1,5 @@
 from flask import json
+
 from tests.data.keyboard_data import (
     keyboard_code_list,
     keyboard_current,
