+++ tests/old/test_os_updater_class.py
