+++ tests/old/test_timezones.py
@@ -1,4 +1,5 @@
 from flask import json
+
 from tests.data.timezone_data import timezones_list
 from tests.utils import dotdict
