+++ tests/test_wifi_manager_class.py
